Start work on new image and embed blocks
