+++ examples/blocks/EmbedBlock.scss
@@ -1,0 +1,41 @@
+.EmbedBlock-container {
+    cursor: pointer;
+}
+
+.EmbedBlock {
+    position: relative;
+}
+
+.EmbedBlock img {
+    max-width: 100%;
+}
+
+.EmbedBlock-icon {
+    position: absolute;
+    top: $controls-spacing;
+    left: $controls-spacing;
+    background: rgba(255, 255, 255, 0.9);
+    line-height: 1;
+    padding-left: $button-spacing;
+    padding-bottom: 0.125rem;
+    pointer-events: none;
+}
+
+.EmbedBlock-preview {
+    max-width: 260px;
+    min-height: 170px;
+    display: inline-block;
+}
+
+.EmbedBlock-options {
+    background: #f3f3f3;
+    border-top: solid 1px #e2e2e2;
+    font-size: 13px;
+    font-family: Open Sans, sans-serif;
+    padding: 1rem;
+
+    h3,
+    h4 {
+        margin-top: 0;
+    }
+}
