+++ examples/blocks/ImageBlock.scss
@@ -1,19 +1,9 @@
-.RichEditor-media {
-    outline: 1px dotted #dadada;
-    position: relative;
-
-    &:hover {
-        outline: 1px dotted $color-teal;
-
-        .RichEditor-media-container {
-            cursor: pointer;
-        }
-    }
+.ImageBlock-container {
+    cursor: pointer;
 }
 
-.RichEditor-media--open {
-    z-index: $activeblock-z-index;
-    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.15);
+.ImageBlock {
+    position: relative;
 }
 
 .RichEditor-grid {
@@ -33,11 +23,11 @@
     box-sizing: border-box;
 }
 
-.RichEditor-media img {
+.ImageBlock img {
     max-width: 100%;
 }
 
-.RichEditor-media-icon {
+.ImageBlock-icon {
     position: absolute;
     top: $controls-spacing;
     left: $controls-spacing;
@@ -48,13 +38,13 @@
     pointer-events: none;
 }
 
-.RichEditor-media-preview {
+.ImageBlock-preview {
     max-width: 260px;
+    min-height: 170px;
     display: inline-block;
-    user-select: none;
 }
 
-.RichEditor-media-options {
+.ImageBlock-options {
     background: #f3f3f3;
     border-top: solid 1px #e2e2e2;
     font-size: 13px;
