+++ examples/blocks/ImageBlock.test.js
@@ -1,21 +1,61 @@
 import React from 'react';
 import { shallow } from 'enzyme';
+import { convertFromRaw, EditorState } from 'draft-js';
+
+import { ENTITY_TYPE } from '../../lib';
+
 import ImageBlock from '../blocks/ImageBlock';
 
+const contentState = convertFromRaw({
+    entityMap: {
+        '0': {
+            type: ENTITY_TYPE.IMAGE,
+            mutability: 'IMMUTABLE',
+            data: {
+                src:
+                    'https://springload.github.io/draftail/static/example-lowres-image.jpg',
+            },
+        },
+    },
+    blocks: [
+        {
+            key: 'b3kdk',
+            text: ' ',
+            type: 'atomic',
+            depth: 0,
+            inlineStyleRanges: [],
+            entityRanges: [
+                {
+                    offset: 0,
+                    length: 1,
+                    key: 0,
+                },
+            ],
+            data: {},
+        },
+    ],
+});
+
 const mockProps = {
-    entity: {
-        getData() {
-            return {
+    block: contentState.getFirstBlock(),
+    blockProps: {
+        editorState: EditorState.createWithContent(contentState),
+        entity: {
+            getType: () => ENTITY_TYPE.IMAGE,
+            getData: () => ({
                 src: 'http://www.example.com/example.png',
                 alt: 'Test alt',
                 alignment: 'right',
-            };
+            }),
         },
+        entityConfig: {
+            type: ENTITY_TYPE.IMAGE,
+        },
+        onChange: jest.fn(),
+        lockEditor: jest.fn(),
+        unlockEditor: jest.fn(),
     },
-    entityConfig: {},
-    isActive: false,
-    onClick: jest.fn(),
-    onSave: jest.fn(),
+    contentState,
 };
 
 describe('ImageBlock', () => {
